feat(lastfm): change scrobble time
Previously, scrobbles were sent as soon as they became acceptable.
However, scrobbles are displayed on last.fm immediately, even if the
track is still displayed as currently playing. This commit changes it to
only submit when the track is finished playing, which displays correctly
on the user's last.fm profile.