refactor: move to tokio (#66)
* chore(nix): update flake.lock

* fix(nix): add rust dev tools to devShell

* fix: gitignore directory specifiers

* chore: add `.vscode/` to gitignore

* refactor(AlbumCache): remove unnecessary mutex

the whole point of moka is to be lock-free

* refactor(AlbumCache): put it in a plain old static

ah this whole thing is silly anyways

* refactor(Decode): remove weird pathbuf -> string shenanigans

* fix: remove some eager allocations

* style: simplify ui::assets::db::load

* chore: remove ahash

* chore: disable `futures` default features

* feat: add tokio

* feat: remove smol

* style: use consistent naming, remove superfluous PlaybackInterface trait

* refactor(QueueItemUIData): use spawn_blocking for lookup and image decode

* fix: replace all spawn(tx.send) with tx.send_blocking

Fixes #61.

just a rough find/replace, so i'd like a quick signoff before merging.

* refactor(last.fm): switch to reqwest from isahc

drops build time by ~40s. should work but want a quick look at this too.

* fix(last.fm): use tokio spawn instead of gpui spawn

* refactor: drop duplicate libs. on linux

* refactor: ditto on macOS

* refactor: ditto on Windows

* fix: playbackcontroller event ordering

move everything to a separate task so events can
be processed sequentially

ultimately i'd like to change the entire
playbackcontroller trait (and most interfaces) to
just a simple recv(pbcevent) fn that each
specialization can handle however it likes, e.g.
serializing to json and handing to a wasm
plugin/whatever, but not trying to boil the ocean
just yet

* fix: scan sends

---------

Co-authored-by: Cameron Ward <cameronuward@gmail.com>